+++ HISTORY.rst
@@ -5,6 +5,8 @@ History
 
 v0.2.0 (TBC)
 ............
+* **breaking change**: ``values()`` on a model is now a method not a property,
+  takes ``include`` and ``exclude`` arguments
 * allow annotation only fields to support mypy
 
 v0.1.0 (2017-06-03)